add logout feature

+++ src/main/java/net/duchung/quora/service/impl/AuthServiceImpl.java
@@ -1,7 +1,9 @@
 package net.duchung.quora.service.impl;
 
 import jakarta.transaction.Transactional;
+import lombok.AllArgsConstructor;
 import net.duchung.quora.common.exception.DataNotFoundException;
+import net.duchung.quora.common.security.JwtBlacklistService;
 import net.duchung.quora.common.security.jwt.JwtUtil;
 import net.duchung.quora.data.entity.User;
 import net.duchung.quora.data.entity.VerificationToken;
@@ -17,41 +19,37 @@ import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.UUID;
 
 @Service
+@AllArgsConstructor
 public class AuthServiceImpl implements AuthService {
-    @Autowired
-    private UserRepository userRepository;
-    @Autowired
-    private PasswordEncoder passwordEncoder;
-    @Autowired
-    private AuthenticationManager authenticationManager;
-    @Autowired
-    private JwtUtil jwtUtil;
-    @Autowired
-    private VerificationTokenRepository verificationTokenRepository;
-
-    @Autowired
-    private MailService mailService;
+    private final UserRepository userRepository;
+    private final PasswordEncoder passwordEncoder;
+    private final AuthenticationManager authenticationManager;
+    private final JwtUtil jwtUtil;
+    private final JwtBlacklistService jwtBlacklistService;
+    private final VerificationTokenRepository verificationTokenRepository;
+    private final MailService mailService;
 
 
     @Override
     public LoginResponse login(String email, String password) {
-        User user = userRepository.findByEmail(email).orElseThrow(() -> new UsernameNotFoundException("Email or password incorrect"));
-        if (!passwordEncoder.matches(password, user.getPassword())) {
-            throw new DataNotFoundException("Email or password incorrect");
-        } else if (!user.isActive()) {
-            throw new DataNotFoundException("User is not active");
-        }
+
         UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(email, password, Collections.emptySet());
-        authenticationManager.authenticate(authenticationToken);
-        return new LoginResponse(jwtUtil.generateToken(user));
+        Authentication authentication = authenticationManager.authenticate(authenticationToken);
+
+        // No need to set SecurityContext because using JWT, a stateless authentication
+//        SecurityContextHolder.getContext().setAuthentication(authentication);
+        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+        return new LoginResponse(jwtUtil.generateToken(userDetails));
     }
 
     @Override
@@ -59,21 +57,18 @@ public class AuthServiceImpl implements AuthService {
     public String register(RegisterRequest registerRequest) {
         String email = registerRequest.getEmail();
         String password = registerRequest.getPassword();
-        String retypePassword = registerRequest.getRetypePassword();
         String fullName = registerRequest.getFullName();
-        if (!password.equals(retypePassword)) {
-            throw new IllegalArgumentException("Passwords do not match");
-        }
         if (userRepository.existsByEmail(email)) {
             throw new DataIntegrityViolationException("Email already exists");
         }
         User user = new User();
         user.setEmail(email);
-        user.setPassword(passwordEncoder.encode(registerRequest.getPassword()));
+        user.setPassword(passwordEncoder.encode(password));
         user.setFullName(fullName);
         User savedUser = userRepository.save(user);
         VerificationToken verificationToken = new VerificationToken(UUID.randomUUID().toString(), savedUser);
         verificationTokenRepository.save(verificationToken);
+        //fixme: send mail maybe fail but user save to db
         mailService.sendVerificationLinkToEmail(email, verificationToken.getToken());
         return "User " + savedUser.getEmail() + " registered successfully! Please check your email for verification.";
     }
@@ -82,19 +77,30 @@ public class AuthServiceImpl implements AuthService {
     public User getCurrentUser() {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         String email = authentication.getName();
-        return userRepository.findByEmail(email).orElseThrow(() -> new UsernameNotFoundException("Email or password incorrect"));
+        return userRepository.findByEmail(email).orElseThrow(() -> new UsernameNotFoundException("There are some problems with your account"));
     }
 
     @Override
+    @Transactional
     public String verify(String code) {
         VerificationToken verificationToken = verificationTokenRepository.findByToken(code);
         if (verificationToken == null) {
             throw new DataNotFoundException("Verification token not found");
+        }
+        if (verificationToken.getExpiryDate().isBefore(LocalDateTime.now())) {
+            throw new DataNotFoundException("Verification token expired");
         }
         User user = verificationToken.getUser();
         user.setActive(true);
         userRepository.save(user);
         verificationTokenRepository.delete(verificationToken);
         return "User " + user.getEmail() + " verified successfully";
+    }
+
+    @Override
+    public String logout(String authHeader) {
+    String token = authHeader.substring(7);
+    jwtBlacklistService.addToBlackList(token);
+    return "Logout success";
     }
 }
